Refactor: MemberTitle package 위치 변경
