+++ src/main/java/ticketaka/mtvs3_final_backend/member/query/service/MemberQueryService.java
@@ -10,8 +10,8 @@ import ticketaka.mtvs3_final_backend.file.query.service.FileQueryService;
 import ticketaka.mtvs3_final_backend.member.command.domain.model.Member;
 import ticketaka.mtvs3_final_backend.member.query.dto.MemberQueryResponseDTO;
 import ticketaka.mtvs3_final_backend.member.query.repository.MemberQueryRepository;
-import ticketaka.mtvs3_final_backend.member.title.command.domain.model.MemberTitle;
-import ticketaka.mtvs3_final_backend.member.title.command.domain.repository.MemberTitleQueryRepository;
+import ticketaka.mtvs3_final_backend.title.member.command.domain.model.MemberTitle;
+import ticketaka.mtvs3_final_backend.title.member.command.domain.repository.MemberTitleQueryRepository;
 import ticketaka.mtvs3_final_backend.sticker.command.domain.model.Sticker;
 import ticketaka.mtvs3_final_backend.sticker.query.service.StickerQueryService;
 import ticketaka.mtvs3_final_backend.ticketing.concert.command.domain.model.Concert;
