+++ src/main/java/ticketaka/mtvs3_final_backend/title/member/command/domain/model/MemberTitle.java
@@ -1,4 +1,4 @@
-package ticketaka.mtvs3_final_backend.member.title.command.domain.model;
+package ticketaka.mtvs3_final_backend.title.member.command.domain.model;
 
 import jakarta.persistence.*;
 import lombok.*;
