+++ src/main/java/ticketaka/mtvs3_final_backend/title/member/command/domain/repository/MemberTitleQueryRepository.java
@@ -1,8 +1,8 @@
-package ticketaka.mtvs3_final_backend.member.title.command.domain.repository;
+package ticketaka.mtvs3_final_backend.title.member.command.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ticketaka.mtvs3_final_backend.member.title.command.domain.model.MemberTitle;
+import ticketaka.mtvs3_final_backend.title.member.command.domain.model.MemberTitle;
 
 import java.util.List;
