+++ src/main/java/ticketaka/mtvs3_final_backend/title/query/service/TitleQueryService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ticketaka.mtvs3_final_backend.member.title.command.domain.model.MemberTitle;
+import ticketaka.mtvs3_final_backend.title.member.command.domain.model.MemberTitle;
 import ticketaka.mtvs3_final_backend.title.command.domain.model.Title;
 import ticketaka.mtvs3_final_backend.title.query.repository.TitleQueryRepository;
